Add support for lock in Update and Delete. MSSQL allow you to specify a lock hint for SELECT, UPDATE and DELETE. To support this, UpdateStatement, DeleteStatement and their manager counterpart need to allow a lock value.
